# src/iron-pets/backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Pets API
  version: 1.0.0
  description: |
    REST API for Iron Pets e-commerce platform - MVP version.
    Provides endpoints for authentication, product catalog, shopping cart, checkout, orders, and user management.
  contact:
    name: Iron Pets Engineering Team
    email: engineering@ironpets.com
  license:
    name: Proprietary
servers:
  - url: https://api.ironpets.com/v1
    description: Production server
  - url: https://staging-api.ironpets.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent if account exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Reset user password using reset token from email
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify/{token}:
    get:
      tags: [Auth]
      summary: Verify email address
      description: Verify user email with verification token
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================================
  # PRODUCT ENDPOINTS
  # ========================================
...
  /products/search:
    get:
      tags: [Products]
      summary: Search products
      description: Full-text search with autocomplete and filters
      operationId: searchProducts
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: brands
          in: query
          schema:
            type: array
            items:
              type: string
        - name: minRating
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating]
      responses:
        '200':
          description: Search results retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
  /cart/merge:
    post:
      tags: [Cart]
      summary: Merge guest cart with user cart
      description: Merge guest cart items into authenticated user cart on login
      operationId: mergeCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeCartRequest'
      responses:
        '200':
          description: Carts merged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

  # ========================================
  # CHECKOUT ENDPOINTS
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 24
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5

    # ========================================
    # AUTH SCHEMAS
    # ========================================
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
        - passwordConfirm
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
        passwordConfirm:
          type: string
          minLength: 8
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            accessToken:
              type: string

    # ========================================
    # USER SCHEMAS
    # ========================================
...
    ProductCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          format: uri
        price:
          type: number
          format: decimal
        comparePrice:
          type: number
          format: decimal
          nullable: true
        averageRating:
          type: number
          minimum: 0
          maximum: 5
        reviewCount:
          type: integer
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        brand:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
...
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productName:
          type: string
        productImage:
          type: string
          format: uri
        productSlug:
          type: string
        price:
          type: number
          format: decimal
        quantity:
          type: integer
          minimum: 1
        lineTotal:
          type: number
          format: decimal
        stockAvailable:
          type: integer
    Cart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: decimal
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CartResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Cart'
...
    MergeCartRequest:
      type: object
      required:
        - guestCartId
      properties:
        guestCartId:
          type: string
          format: uuid

    # ========================================
    # CHECKOUT SCHEMAS
    # ========================================
...